data/parser/requirements.txt
pandas
python-dateutil
cx_Oracle
